+++ src/App.js
@@ -2,9 +2,9 @@ import React from "react";
 
 import Dashboard from "components/Dashboard";
 import LandingPage from "components/LandingPage/LandingPage";
-import LoginPage from "components/LoginPage";
+import LoginPage from "components/LoginPage/LoginPage";
 import ProtectedRoutes from "components/ProtectedRoutes/ProtectedRoutes";
-import SignupPage from "components/SignupPage/supage";
+import SignupPage from "components/SignupPage/SignupPage";
 import { useDispatch, useSelector } from "react-redux";
 import { Navigate, Route, Routes, useNavigate } from "react-router-dom";
 import { initializeLoggedInUser } from "reducers/loggedInUserReducer.js";
